[wip] refactor code to separate webkit and chromium better
* splash.engines.webkit and splash.engines.chromium
* added splash.log
* splash.exceptions is renamed to splash.errors; RenderErrorInfo is moved here
