+++ splash/engines/webkit/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from .browser_tab import WebkitBrowserTab
