+++ splash/engines/webkit/webpage.py
@@ -1,35 +1,17 @@
 # -*- coding: utf-8 -*-
-from collections import namedtuple
 
 import sip
-from PyQt5.QtWebKitWidgets import QWebPage, QWebView
+from PyQt5.QtWebKitWidgets import QWebPage
 from PyQt5.QtCore import QByteArray
 from twisted.python import log
 import traceback
 
+from splash.browser_tab import WebpageEventLogger
 from splash.har_builder import HarBuilder
-
-RenderErrorInfo = namedtuple('RenderErrorInfo', 'type code text url')
-
-
-class SplashQWebView(QWebView):
-    """
-    QWebView subclass that handles 'close' requests.
-    """
-    onBeforeClose = None
-
-    def closeEvent(self, event):
-        dont_close = False
-        if self.onBeforeClose:
-            dont_close = self.onBeforeClose()
-
-        if dont_close:
-            event.ignore()
-        else:
-            event.accept()
+from splash.errors import RenderErrorInfo
 
 
-class SplashQWebPage(QWebPage):
+class WebkitWebPage(QWebPage):
     """
     QWebPage subclass that:
 
@@ -107,7 +89,7 @@ class SplashQWebPage(QWebPage):
             self.run_callbacks('on_navigation_locked', networkRequest)
             return False
         self.error_info = None
-        return super(SplashQWebPage, self).acceptNavigationRequest(webFrame, networkRequest, navigationType)
+        return super(WebkitWebPage, self).acceptNavigationRequest(webFrame, networkRequest, navigationType)
 
     def javaScriptAlert(self, frame, msg):
         return
@@ -127,7 +109,7 @@ class SplashQWebPage(QWebPage):
 
     def userAgentForUrl(self, url):
         if self.custom_user_agent is None:
-            return super(SplashQWebPage, self).userAgentForUrl(url)
+            return super(WebkitWebPage, self).userAgentForUrl(url)
         else:
             return self.custom_user_agent
 
@@ -200,3 +182,43 @@ class SplashQWebPage(QWebPage):
 
     def error_loading(self, load_finished_ok):
         return load_finished_ok and self.error_info is not None
+
+
+class WebkitEventLogger(WebpageEventLogger):
+    """ This class logs various events that happen with QWebPage """
+    def add_web_page(self, web_page: WebkitWebPage) -> None:
+        frame = web_page.mainFrame()
+        # setup logging
+        if self.logger.verbosity >= 4:
+            web_page.loadStarted.connect(self.on_load_started)
+            frame.loadFinished.connect(self.on_frame_load_finished)
+            frame.loadStarted.connect(self.on_frame_load_started)
+            frame.contentsSizeChanged.connect(self.on_contents_size_changed)
+            # TODO: on_repaint
+
+        if self.logger.verbosity >= 3:
+            frame.javaScriptWindowObjectCleared.connect(self.on_javascript_window_object_cleared)
+            frame.initialLayoutCompleted.connect(self.on_initial_layout_completed)
+            frame.urlChanged.connect(self.on_url_changed)
+
+    def on_load_started(self):
+        self.logger.log("loadStarted")
+
+    def on_frame_load_finished(self, ok):
+        self.logger.log("mainFrame().LoadFinished %s" % ok)
+
+    def on_frame_load_started(self):
+        self.logger.log("mainFrame().loadStarted")
+
+    def on_contents_size_changed(self, sz):
+        self.logger.log("mainFrame().contentsSizeChanged: %s" % sz)
+
+    def on_javascript_window_object_cleared(self):
+        self.logger.log("mainFrame().javaScriptWindowObjectCleared")
+
+    def on_initial_layout_completed(self):
+        self.logger.log("mainFrame().initialLayoutCompleted")
+
+    def on_url_changed(self, url):
+        self.logger.log("mainFrame().urlChanged %s" % qurl2ascii(url))
+
